requirements.txt

---

### 2. `requirements.txt` ✔️
Para instalar las dependencias con `pip`:
```txt
openai-whisper
fuzzywuzzy
python-Levenshtein
torch
